requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-multipart==0.0.6
jinja2==3.1.3
pydantic==2.6.0
pydantic-settings==2.1.0

# Qdrant
qdrant-client==1.7.2

# Embeddings
sentence-transformers==2.3.1
torch==2.2.0

# LLM
groq==0.4.2

# LangChain Stack
langchain>=0.2.0,<0.3.0
langchain-core>=0.2.27,<0.3.0
langgraph>=0.2.0,<0.3.0

# Search
rank-bm25==0.2.2

# Utils
numpy==1.26.3
tqdm==4.66.1
python-dotenv==1.0.0
-e .
